Modify DNS resolution.
I am not interested in the IPs that names resolve to. I just want to know that a new cert
has been issued and that it has a valid piece of DNS data.

There is no point in querying for A and CNAME records as the stub resolver will
follow up an answer of CNAME with an A record query.

This is just duplicating work.