Remove QNAME to QTYPE mapping.
I do not want this mapping as I am only interested in if the CN of the new cert
resolves to something publicly accessible on the internet.

Also there is a bunch of casting of data types going on from dict to set to list.

If we want unique names then just use a set from the start.