Add backoff to crt.sh queries.
There are two modes in this script for discovering newly issued certs.

One will connect to a DBMS and issue SQL queries. The other will issue HTTP requests
to an API.

The HTTP requests were using the same static timeout meaning if you had a result set with a large
number of entries. The read() operation on the socket would timeout.

This now uses exponential backoff.