Add async DNS and HTTP requests.
There is still a lot of code I don't understand. So I am writing this in a way
that it modifies the existing structure of the code as little as possible.

DNS resolution and HTTP requests are now async and the only reported domains
are URLs that we can actually visit over HTTP or HTTPs.