Track new URLs using database.
Currently the code uses text files in the output directory to store new
URLs. It then uses these files to track new URLs between subsequent
runs.

This is fine but the code is a little messy and complicated so I am
replacing it with a very basic SQLite DB.

I am purposely not touching the code for manipulating the text files
right now as I will likely break something due to the lack of tests this
code base has.

I am using the new SQLite DB code to do the new URL diff check in the
function that is used for posting results to slack.