Emulate iOS device, add page_title and input form check, pass Host: header.
There are several changes here.

I have added some headers to try simulate a real device just in case we were getting blocked by things like
cloudflare.

I am now extracting the page_title and doing a little check to see if I can detect input forms on the page so
these can be prioritized when searching the results.

It is a very basic regular expression so it is definitely going to have edge cases and failures. It would be best
to use lxml or bs4 to actually parse the HTML content and traverse the DOM tree.

Lastly I realised if we hit a load balancer it is going to struggle to know which website we are requesting because'
I am first doing the DNS resolution of the hostname to grab the IP and then visiting the IP over HTTP/HTTPs.

If we hit a load balancer that has multiple tenants then it isn't going to know which website it should request as
the host header is missing.

So now the host header is set to the hostname we resolved in the DNS resolution function.