Remove cert_database class.
From looking at the code I can not really see a reason why this needs to be a class. A
single instance is created so at most it could be a singleton. Using a class just to have dot
notation is complete overkill.

I also made error logging on by default. I want to know about errors and right now the config
is in a piece of python code. I will be moving it to a config.toml file eventually and adding
actual python logging.