Stage changes to switch over to TOML config.
Adding some of the skeleton code for switching over to use TOML as configuration
rather than having configuration defined in code.

The little config.py will only give back a single instance of the config each time it
is called.

I am planning on breaking out the code in sublert.py into the lib/ directory
so this will be helpful then.