Return none instead of exiting when parsing -u CLI option.
This was causing the script to exit after doing one loop. I thought I was
returning None here but I was actually exiting.

Now that it returns none when you do not pass a domain to add via the -u
option the code will properly loop using the applied schedule.