Use cron to manage schedule.
There was a bug in the python native scheduling that I have yet to figure out
the root cause of. I suspect it is to do with the async code and scheduling at the
same time.

There is likely some syncing issues with the different threads.

For now lets just revert back to using good old cron with some scaffolding
around setting up the dependencies and crontab.