Remove querying crt.sh over HTTP.
Having the two methods to query crt.sh introduces a modality in the way
that the lookup() function can fail.

I also noticed when running/testing the code that it rarely fallsback
to doing the query over HTTP and when it does it frequently times out
even with the backoff code.

So I am just going to remove it and fail fast if looking up the names using
SQL queries returns an empty result.