Add jitter.
The run.sh script is invoked via cron. I want to add a jitter so I am adding
it in the script rather than touching the crontab file.

The queries to the crt.sh database over SQL have no jitter either so I should
update that too so their database is not being hammered.

Right now a query per domain in domains.txt is sent so if you have 1000 domains
in domains.txt the code will try to run 1000 concurrent queries.